+++ database/migrations/2026_03_05_035000_create_ai_infrastructure_tables.php
@@ -14,7 +14,7 @@ return new class extends Migration
         if (!Schema::hasTable('ai_audit_logs')) {
             Schema::create('ai_audit_logs', function (Blueprint $table) {
                 $table->id();
-                $table->unsignedInteger('user_id')->nullable();
+                $table->unsignedBigInteger('user_id')->nullable();
                 $table->string('feature', 50); // assistant, listing_gen, photo_enhance, anti_scam, reputation
                 $table->string('provider', 20); // openai, gemini, anthropic
                 $table->string('model', 50)->nullable();
@@ -37,7 +37,7 @@ return new class extends Migration
         if (!Schema::hasTable('scam_signals')) {
             Schema::create('scam_signals', function (Blueprint $table) {
                 $table->id();
-                $table->unsignedInteger('flagged_user_id')->nullable();
+                $table->unsignedBigInteger('flagged_user_id')->nullable();
                 $table->unsignedBigInteger('flagged_listing_id')->nullable();
                 $table->string('signal_type', 50); // duplicate_listing, suspicious_price, forbidden_words
                 $table->float('probability')->default(0);
@@ -55,7 +55,7 @@ return new class extends Migration
         if (!Schema::hasTable('reputation_scores')) {
             Schema::create('reputation_scores', function (Blueprint $table) {
                 $table->id();
-                $table->unsignedInteger('user_id')->unique();
+                $table->unsignedBigInteger('user_id')->unique();
                 $table->integer('points')->default(100);
                 $table->string('level_badge')->default('Newcomer');
                 $table->float('trust_score')->default(5.0);
